fix: migrate config to new v41 schema (#131)
Update the Renovate app config with the new schema changes.

Remove the Yoast host rules as this plugin is no longer used by Articles.

This also removes the config validation workflow as it is not meant to be used to validate a 
self-hosted Renovate app's config and is now failing with the schema change to use 
allowedCommands. It was just "luck" that it was working before.

The validation workflow in renovate-config repo will be kept in place and continue to 
 Renovate rule configuration errors.